+++ packages/terra-form-select/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added visual focus dashed border for `terra-form-select` tags.
+
 * Fixed
   * Fixed accessibility issue in `MultiSelect` component.